+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- upcloud_kubernetes_cluster: add ephemeral resource for getting kubeconfig of a cluster.
+
 ## [5.33.4] - 2026-02-04
 
 ### Fixed